requirements.txt
discord.py
-e git+https://github.com/TerryRPatterson/discord_bot_framework#egg=discord-bot-framework-terryp
mongoengine
python-dateutil
